vertex_ai/trigger_cf/requirements.txt
kfp==1.8.14
google-cloud-aiplatform==1.18.2
google-cloud-storage



